+++ src/index.js
@@ -1,5 +1,5 @@
-import Treebeard from './components/treebeard';
-import decorators from './components/decorators';
+import Treebeard from './components';
+import decorators from './components/Decorators';
 import animations from './themes/animations';
 import theme from './themes/default';
